added: BookView

+++ book_app/models.py
@@ -5,4 +5,4 @@ from django.db import models
 class Books(models.Model):
     title = models.CharField(max_length=50, default="No title defined")
     author = models.CharField(max_length=25, default="No author defined")
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=8, decimal_places=2)
